requirements.txt
pandas==2.3.3
numpy==2.3.5
openai==2.9.0
python-dotenv==1.2.1
requests==2.31.0
google-genai==0.6.5